requirements.txt
google-genai
python-dotenv
python-telegram-bot
beautifulsoup4
playwright